+++ airdnb/src/main/resources/static/css/accommodation/createForm.css
@@ -1,33 +1,29 @@
-/* 기본 스타일 */
+/* createForm.css */
+
 body {
-    font-family: 'Arial', sans-serif;
-    background-color: #f7f7f7;
-    margin: 0;
-    padding: 0;
+    font-family: Arial, sans-serif;
+    margin: 20px;
     display: flex;
-    justify-content: center;
-    align-items: center;
-    height: 100vh;
+    justify-content: center; /* Center align the content horizontally */
+    align-items: center; /* Center align the content vertically */
+    flex-direction: column; /* Align items in a column */
 }
 
-/* 컨테이너 스타일 */
 .container {
     background-color: #fff;
     padding: 20px 30px;
     border-radius: 10px;
     box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
-    max-width: 500px;
     width: 100%;
+    max-width: 660px;
 }
 
-/* 제목 스타일 */
 h1 {
     text-align: center;
     color: #333;
     margin-bottom: 20px;
 }
 
-/* 라벨 스타일 */
 label {
     display: block;
     margin-bottom: 5px;
@@ -35,7 +31,6 @@ label {
     font-weight: bold;
 }
 
-/* 입력 필드 및 셀렉트 스타일 */
 input, select {
     width: 100%;
     padding: 10px;
@@ -46,14 +41,12 @@ input, select {
     transition: border-color 0.3s;
 }
 
-/* 입력 필드 및 셀렉트 포커스 스타일 */
 input:focus, select:focus {
     border-color: #007bff;
     outline: none;
 }
 
-/* 버튼 스타일 */
-button {
+.submit_btn {
     width: 100%;
     padding: 12px;
     background-color: #007bff;
@@ -65,12 +58,10 @@ button {
     transition: background-color 0.3s;
 }
 
-/* 버튼 호버 스타일 */
-button:hover {
+.submit_btn:hover {
     background-color: #0056b3;
 }
 
-/* 에러 메시지 스타일 */
 .error-message {
     color: red;
     font-size: 0.9em;
@@ -79,7 +70,52 @@ button:hover {
     display: none;
 }
 
-/* 에러 메시지 활성화 스타일 */
 input:invalid, select:invalid {
     border-color: red;
+}
+
+/* 새로운 체크박스 스타일 */
+#amenities {
+    display: flex;
+    flex-wrap: wrap;
+    gap: 10px;
+}
+
+#amenities label {
+    display: flex;
+    align-items: center;
+    font-weight: normal;
+    margin-bottom: 0;
+}
+
+#amenities input[type="checkbox"] {
+    appearance: none;
+    width: 20px;
+    height: 20px;
+    border: 2px solid #007bff;
+    border-radius: 3px;
+    margin-right: 5px;
+    transition: background-color 0.3s, border-color 0.3s;
+    cursor: pointer;
+    display: inline-block;
+    position: relative;
+}
+
+#amenities input[type="checkbox"]:checked {
+    background-color: #007bff;
+    border-color: #007bff;
+}
+
+#amenities input[type="checkbox"]:checked::before {
+    content: '✔';
+    display: block;
+    color: white;
+    font-weight: bold;
+    text-align: center;
+    line-height: 18px;
+    position: absolute;
+    top: 0;
+    left: 0;
+    width: 100%;
+    height: 100%;
 }
