+++ airdnb/src/main/resources/static/css/setting/top-btn.css
@@ -1,10 +1,10 @@
-/* 버튼 컨테이너 스타일 */
+/* top-btn.css */
+
 .top-button-container {
     display: flex;
     justify-content: flex-start;
     align-items: center;
     margin-bottom: 20px;
-    padding-left: 20px; /* 왼쪽 정렬을 위해 여백 추가 */
 }
 
 /* 버튼 스타일 */
