+++ airdnb/src/main/resources/static/js/setting/top-btn.js
@@ -4,7 +4,8 @@ document.addEventListener("DOMContentLoaded", () => {
     const buttons = [
         { text: '숙소 생성 페이지', url: '/accommodation/create' },
         { text: '숙소 유형 페이지', url: '/accommodation-type' },
-        { text: '숙소 방 유형 페이지', url: '/accommodation-room-type' }
+        { text: '숙소 방 유형 페이지', url: '/accommodation-room-type' },
+        { text: '숙소 편의시설 페이지', url: '/amenity' } // New button added
     ];
 
     buttons.forEach(button => {
