+++ airdnb/src/main/resources/templates/amenity/main.html
@@ -1,0 +1,22 @@
+<!DOCTYPE html>
+<html>
+<head>
+    <title>Amenities Management</title>
+    <link rel="stylesheet" href="/css/amenity/main.css">
+    <link rel="stylesheet" href="/css/setting/top-btn.css">
+</head>
+<body>
+<div class="top-button-container"></div>
+<h1>숙소 편의시설 페이지</h1>
+<button class="create-button">Create</button>
+<div class="create-form" id="createForm">
+    <input type="text" id="amenityName" placeholder="편의시설 이름">
+    <button class="save-button">Save</button>
+</div>
+<ul id="amenityList">
+    <!-- js data insert area -->
+</ul>
+<script src="/js/amenity/main.js"></script>
+<script src="/js/setting/top-btn.js"></script>
+</body>
+</html>
